Sent statementSequence transformer to utils + transform happens immediately before execution
